factor our I/O access
